refactored to pass maven tests
