--- web/src/main/webapp/WEB-INF/si-config-measurement.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:int="http://www.springframework.org/schema/integration"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:int-stream="http://www.springframework.org/schema/integration/stream"
-       xmlns:int-file="http://www.springframework.org/schema/integration/file"
-       xmlns:int-http="http://www.springframework.org/schema/integration/http"
-       xsi:schemaLocation="
-        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-        http://www.springframework.org/schema/integration http://www.springframework.org/schema/integration/spring-integration.xsd
-        http://www.springframework.org/schema/integration/stream http://www.springframework.org/schema/integration/stream/spring-integration-stream.xsd
-        http://www.springframework.org/schema/integration/http http://www.springframework.org/schema/integration/http/spring-integration-http-4.3.xsd
-        http://www.springframework.org/schema/integration/file http://www.springframework.org/schema/integration/file/spring-integration-file.xsd
-        http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd
-        ">
-
-    <context:property-placeholder
-            properties-ref="systemProperties"/>
-    <!-- Properties Bean linux: git/ilona/ilona/src/main/webapp/WEB-INF/mybatis-configuration.xml -->
-    <!-- Properties Bean windows: mybatis-configuration.xml (default! desktop) -->
-    <bean id="systemProperties"
-          class="org.springframework.beans.factory.config.PropertiesFactoryBean">
-        <property name="properties">
-            <props>
-                <prop key="measurement.host">${measurement.host}</prop>
-                <prop key="measurement.port">${measurement.port}</prop>
-            </props>
-        </property>
-    </bean>
-
-    <int:gateway
-            service-interface="uni.miskolc.ips.ilona.positioning.service.gateway.ZoneQueryService"
-            id="ZoneGateway" default-request-channel="ZoneReqChannel"/>
-
-    <int:gateway
-            service-interface="uni.miskolc.ips.ilona.positioning.service.gateway.MeasurementGateway"
-            id="MeasurementGateway" default-request-channel="MeasurementReqChannel"/>
-
-
-
-    <int:channel id="ZoneReqChannel"/>
-
-    <int:channel id="MeasurementReqChannel"/>
-
-
-
-
-    <int-http:outbound-gateway  request-channel="ZoneReqChannel"
-                                url="http://${measurement.host}:${measurement.port}/ilona/listZones"
-                                http-method="GET"
-                                extract-request-payload="true"
-                                expected-response-type="uni.miskolc.ips.ilona.measurement.model.position.Zone[]">
-    </int-http:outbound-gateway>
-
-
-    <int-http:outbound-gateway  request-channel="MeasurementReqChannel"
-                                url="http://${measurement.host}:${measurement.port}/ilona/resources/listMeasurements"
-                                http-method="GET"
-                                extract-request-payload="true"
-                                expected-response-type="uni.miskolc.ips.ilona.measurement.model.measurement.Measurement[]">
-    </int-http:outbound-gateway>
-
-
-</beans>
